requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.2
requests==2.33.1
ruff==0.15.9
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
flask>=3.0
supabase>=2.0
python-dotenv>=1.0
openai>=1.0